Create multi-project for rediscala
Benchmarks can be executed using `sbt bench/test` and don't block the
root project and its tests anymore.
