+++ src/bench/src/test/scala/rediscala/benchmark/RedisBench.scala
@@ -1,4 +1,4 @@
-package bench
+package rediscala.benchmark
 
 import scala.concurrent._
 import scala.concurrent.duration._
@@ -7,14 +7,13 @@ import akka.actor.ActorSystem
 import scala.collection.Iterator
 
 import org.scalameter._
-import redis.{RedisServer, RedisClientPool, RedisClient}
-import org.scalameter.execution
-
 import org.scalameter.api.{Executor,Aggregator,Gen,Reporter,RegressionReporter,HtmlReporter,SerializationPersistor}
+import redis.RedisClient
+import org.scalameter.execution
 
 import org.scalameter.picklers.noPickler._
 
-object RedisBenchPool extends Bench[Double] {
+object RedisBench extends Bench[Double] {
 
   override def reporter: Reporter[Double] = Reporter.Composite(
     new RegressionReporter[Double](
@@ -31,22 +30,19 @@ object RedisBenchPool extends Bench[Double] {
     def numeric: Numeric[Double] = implicitly[Numeric[Double]]
   }
 
-  //def measurer: Measurer = new Executor.Measurer.MemoryFootprint
-
   def executor: Executor[Double] = new execution.SeparateJvmsExecutor(warmer, aggregator, measurer)
-
 
   def persistor = new SerializationPersistor()
 
-  def exponential(axisName: String)(from: Int, until: Int, factor: Int): Gen[(Int, RedisBenchContextPool)] = new Gen[(Int, RedisBenchContextPool)] {
+  def exponential(axisName: String)(from: Int, until: Int, factor: Int): Gen[(Int, RedisBenchContext)] = new Gen[(Int, RedisBenchContext)] {
     def warmupset = {
-      Iterator.single(((until - from) / 2, new RedisBenchContextPool()))
+      Iterator.single(((until - from) / 2, new RedisBenchContext()))
     }
 
     def dataset = Iterator.iterate(from)(_ * factor).takeWhile(_ <= until).map(x => Parameters(new Parameter[String](axisName) -> x))
 
     def generate(params: Parameters) = {
-      (params[Int](axisName), new RedisBenchContextPool())
+      (params[Int](axisName), new RedisBenchContext())
     }
   }
 
@@ -59,7 +55,7 @@ object RedisBenchPool extends Bench[Double] {
       using(sizes).setUp(redisSetUp())
         .tearDown(redisTearDown)
         .in {
-        case (i: Int, redisBench: RedisBenchContextPool) =>
+        case (i: Int, redisBench: RedisBenchContext) =>
           val redis = redisBench.redis
           implicit val ec = redis.executionContext
 
@@ -77,7 +73,7 @@ object RedisBenchPool extends Bench[Double] {
       using(sizes).setUp(redisSetUp())
         .tearDown(redisTearDown)
         .in {
-        case (i: Int, redisBench: RedisBenchContextPool) =>
+        case (i: Int, redisBench: RedisBenchContext) =>
           val redis = redisBench.redis
           implicit val ec = redis.executionContext
 
@@ -95,14 +91,14 @@ object RedisBenchPool extends Bench[Double] {
       using(sizes).setUp(redisSetUp(_.set("a", "abc")))
         .tearDown(redisTearDown)
         .in {
-        case (i: Int, redisBench: RedisBenchContextPool) =>
+        case (i: Int, redisBench: RedisBenchContext) =>
           val redis = redisBench.redis
           implicit val ec = redis.executionContext
 
           val r = for {
             ii <- 0 until i
           } yield {
-            redis.get("a")
+            redis.get("i")
           }
           Await.result(Future.sequence(r), 30 seconds)
       }
@@ -110,19 +106,20 @@ object RedisBenchPool extends Bench[Double] {
 
   }
 
-  def redisSetUp(init: RedisClient => Unit = _ => {})(data: (Int, RedisBenchContextPool)) = data match {
-    case (i: Int, redisBench: RedisBenchContextPool) =>
+  def redisSetUp(init: RedisClient => Unit = _ => {})(data: (Int, RedisBenchContext)) = data match {
+    case (i: Int, redisBench: RedisBenchContext) => {
       redisBench.akkaSystem = akka.actor.ActorSystem()
-      redisBench.redis = RedisClientPool(Seq(RedisServer(), RedisServer(), RedisServer()))(redisBench.akkaSystem)
+      redisBench.redis = RedisClient()(redisBench.akkaSystem)
       Await.result(redisBench.redis.ping(), 2 seconds)
+    }
   }
 
-  def redisTearDown(data: (Int, RedisBenchContextPool)) = data match {
-    case (i: Int, redisBench: RedisBenchContextPool) =>
+  def redisTearDown(data: (Int, RedisBenchContext)) = data match {
+    case (i: Int, redisBench: RedisBenchContext) =>
       redisBench.redis.stop()
       redisBench.akkaSystem.terminate()
       Await.result(redisBench.akkaSystem.whenTerminated, Duration.Inf)
   }
 }
 
-class RedisBenchContextPool(var redis: RedisClientPool = null, var akkaSystem: ActorSystem = null)
+class RedisBenchContext(var redis: RedisClient = null, var akkaSystem: ActorSystem = null)
