+++ src/bench/src/test/scala/rediscala/benchmark/RedisBenchPool.scala
@@ -1,4 +1,5 @@
-package bench
+package rediscala.benchmark
+
 
 import scala.concurrent._
 import scala.concurrent.duration._
@@ -7,13 +8,14 @@ import akka.actor.ActorSystem
 import scala.collection.Iterator
 
 import org.scalameter._
-import org.scalameter.api.{Executor,Aggregator,Gen,Reporter,RegressionReporter,HtmlReporter,SerializationPersistor}
-import redis.RedisClient
+import redis.{RedisServer, RedisClientPool, RedisClient}
 import org.scalameter.execution
+
+import org.scalameter.api.{Executor,Aggregator,Gen,Reporter,RegressionReporter,HtmlReporter,SerializationPersistor}
 
 import org.scalameter.picklers.noPickler._
 
-object RedisBench extends Bench[Double] {
+object RedisBenchPool extends Bench[Double] {
 
   override def reporter: Reporter[Double] = Reporter.Composite(
     new RegressionReporter[Double](
@@ -30,19 +32,22 @@ object RedisBench extends Bench[Double] {
     def numeric: Numeric[Double] = implicitly[Numeric[Double]]
   }
 
+  //def measurer: Measurer = new Executor.Measurer.MemoryFootprint
+
   def executor: Executor[Double] = new execution.SeparateJvmsExecutor(warmer, aggregator, measurer)
+
 
   def persistor = new SerializationPersistor()
 
-  def exponential(axisName: String)(from: Int, until: Int, factor: Int): Gen[(Int, RedisBenchContext)] = new Gen[(Int, RedisBenchContext)] {
+  def exponential(axisName: String)(from: Int, until: Int, factor: Int): Gen[(Int, RedisBenchContextPool)] = new Gen[(Int, RedisBenchContextPool)] {
     def warmupset = {
-      Iterator.single(((until - from) / 2, new RedisBenchContext()))
+      Iterator.single(((until - from) / 2, new RedisBenchContextPool()))
     }
 
     def dataset = Iterator.iterate(from)(_ * factor).takeWhile(_ <= until).map(x => Parameters(new Parameter[String](axisName) -> x))
 
     def generate(params: Parameters) = {
-      (params[Int](axisName), new RedisBenchContext())
+      (params[Int](axisName), new RedisBenchContextPool())
     }
   }
 
@@ -55,7 +60,7 @@ object RedisBench extends Bench[Double] {
       using(sizes).setUp(redisSetUp())
         .tearDown(redisTearDown)
         .in {
-        case (i: Int, redisBench: RedisBenchContext) =>
+        case (i: Int, redisBench: RedisBenchContextPool) =>
           val redis = redisBench.redis
           implicit val ec = redis.executionContext
 
@@ -73,7 +78,7 @@ object RedisBench extends Bench[Double] {
       using(sizes).setUp(redisSetUp())
         .tearDown(redisTearDown)
         .in {
-        case (i: Int, redisBench: RedisBenchContext) =>
+        case (i: Int, redisBench: RedisBenchContextPool) =>
           val redis = redisBench.redis
           implicit val ec = redis.executionContext
 
@@ -91,14 +96,14 @@ object RedisBench extends Bench[Double] {
       using(sizes).setUp(redisSetUp(_.set("a", "abc")))
         .tearDown(redisTearDown)
         .in {
-        case (i: Int, redisBench: RedisBenchContext) =>
+        case (i: Int, redisBench: RedisBenchContextPool) =>
           val redis = redisBench.redis
           implicit val ec = redis.executionContext
 
           val r = for {
             ii <- 0 until i
           } yield {
-            redis.get("i")
+            redis.get("a")
           }
           Await.result(Future.sequence(r), 30 seconds)
       }
@@ -106,20 +111,19 @@ object RedisBench extends Bench[Double] {
 
   }
 
-  def redisSetUp(init: RedisClient => Unit = _ => {})(data: (Int, RedisBenchContext)) = data match {
-    case (i: Int, redisBench: RedisBenchContext) => {
+  def redisSetUp(init: RedisClient => Unit = _ => {})(data: (Int, RedisBenchContextPool)) = data match {
+    case (i: Int, redisBench: RedisBenchContextPool) =>
       redisBench.akkaSystem = akka.actor.ActorSystem()
-      redisBench.redis = RedisClient()(redisBench.akkaSystem)
+      redisBench.redis = RedisClientPool(Seq(RedisServer(), RedisServer(), RedisServer()))(redisBench.akkaSystem)
       Await.result(redisBench.redis.ping(), 2 seconds)
-    }
   }
 
-  def redisTearDown(data: (Int, RedisBenchContext)) = data match {
-    case (i: Int, redisBench: RedisBenchContext) =>
+  def redisTearDown(data: (Int, RedisBenchContextPool)) = data match {
+    case (i: Int, redisBench: RedisBenchContextPool) =>
       redisBench.redis.stop()
       redisBench.akkaSystem.terminate()
       Await.result(redisBench.akkaSystem.whenTerminated, Duration.Inf)
   }
 }
 
-class RedisBenchContext(var redis: RedisClient = null, var akkaSystem: ActorSystem = null)
+class RedisBenchContextPool(var redis: RedisClientPool = null, var akkaSystem: ActorSystem = null)
